Separate api routes to data.js
